visual-tree-search-backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
websockets>=11.0.3
python-dotenv>=1.0.0
supabase>=1.0.3
httpx>=0.24.1
openai==1.42.0
playwright==1.39.0
watchdog==4.0.2
pyparsing==3.1.2
litellm==1.42.1
Pillow>=10.1
beautifulsoup4>=4.12
elevenlabs==1.6.1
numpy>= 1.14
lxml>= 4.9
httpx==0.27.2
evaluate==0.4.0
beartype==0.12.0
scikit-image==0.22.0
numpy==1.25.2
aiolimiter==1.1.0
transformers==4.34.0
nltk==3.8.1